refactor authentication module
